requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
python-multipart
pillow
pydantic[email]
fastapi-mail
cloudinary
httpx==0.27.0
cryptography==46.0.3
python-dotenv==1.0.0
alembic
setuptools